LayoutTests/imported/w3c/web-platform-tests/navigation-api/currententrychange-event/navigation-navigate-replace-same-doc-expected.txt


FAIL currententrychange fires for navigation.navigate() with replace assert_equals: expected -1 but got 1
